Include player name in achievement popups during team games
